PYIC-3796: Move all validation from lambda handler class to criCheckingService
